Optimizing types.

+++ src/shared/buffer.c
@@ -1,7 +1,7 @@
 #include "buffer.h"
 #include <stdio.h>
 
-uint16_t pack_nw_buff(struct NWBuffer* buffer, void* new_data, uint16_t new_data_len) {
+uint16_t pack_nw_buff(struct NWBuffer* buffer, void* new_data, size_t new_data_len) {
   if(buffer->length + new_data_len > NW_BUF_LEN) {
     // Buffer full.
     return 0;
@@ -21,10 +21,10 @@ uint16_t pack_nw_buff(struct NWBuffer* buffer, void* new_data, uint16_t new_data
  * Updates offset value to point to next struct if required. */
 void* unpack_nw_buff(
     struct NWBuffer* buffer,
-    uint16_t payload_offset,
-    uint16_t* new_payload_offset,   // May be null if final offset not needed.
+    size_t payload_offset,
+    size_t* new_payload_offset,   // May be null if final offset not needed.
     void* dest_container,           // May be null if no copy to struct needed.
-    uint16_t dest_container_len
+    size_t dest_container_len
 ) {
   if(payload_offset + dest_container_len > buffer->length) {
     // Requested data overlaps end of buffer.
